+++ Sources/DMS/v3/APIs/Projects/DMSProjectsRequestFiltersV3.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+public extension DMS {
+    
+    enum ProjectsRequestFilterOptionV3: FilterOption {
+        
+        /// IDs of projects
+        case id(_ value: [Int])
+        case name(_ value: String)
+        case shortName(_ value: String)
+        case address(_ value: String)
+        case country(_ value: String)
+    }
+    
+    class ProjectsRequestFiltersV3: RequestFilters<ProjectsRequestFilterOptionV3> {}
+}