+++ Sources/DMS/v3/APIs/Users/DMSUsersApiV3.swift
@@ -1,0 +1,67 @@
+import Foundation
+import DLJSONAPI
+
+public extension DMS {
+    
+    class UsersApiV3: JSONAPI.BaseApi {
+        
+        // MARK: - Public properties
+        
+        public let requestBuilder: UsersRequestBuilderV3
+        
+        // MARK: -
+        
+        public required init(apiStack: JSONAPI.BaseApiStack) {
+            self.requestBuilder = .init(
+                builderStack: .fromApiStack(apiStack)
+            )
+            
+            super.init(apiStack: apiStack)
+        }
+    }
+}
+
+// MARK: - Public methods
+
+public extension DMS.UsersApiV3 {
+    
+    @discardableResult
+    func getUser(
+        for accountId: String,
+        include: [String]?,
+        completion: @escaping ((_ result: RequestSingleResult<DMS.UserResource>) -> Void)
+    ) -> Cancelable {
+        
+        let cancelable = self.network.getEmptyCancelable()
+        
+        requestBuilder.buildGetUser(
+            for: accountId,
+            include: include,
+            completion: { [weak self] (request) in
+                
+                guard let request = request else {
+                    completion(.failure(JSONAPIError.failedToSignRequest))
+                    return
+                }
+                
+                cancelable.cancelable = self?.requestSingle(
+                    DMS.UserResource.self,
+                    request: request,
+                    completion: { (result) in
+                        
+                        switch result {
+                        
+                        case .failure(let error):
+                            completion(.failure(error))
+                            
+                        case .success(let document):
+                            completion(.success(document))
+                        }
+                    }
+                )
+            }
+        )
+        
+        return cancelable
+    }
+}